% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_epoch_col.R
\name{add_epoch_col}
\alias{add_epoch_col}
\title{ruf}
\usage{
add_epoch_col(data, every_n_rows = 10)
}
\arguments{
\item{data}{class: dataframe}

\item{every_n_rows}{class: numeric}
}
\description{
ruf
}
\examples{
add_epoch_col(df)
}
\keyword{data}
\keyword{processing}
