% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_harmonization_cols.R
\name{add_harmonization_columns}
\alias{add_harmonization_columns}
\title{ruf}
\usage{
add_harmonization_columns(harmonization_table, stacked_dataset)
}
\arguments{
\item{harmonization_table}{class: tibble}

\item{stacked_dataset}{class: tibble}
}
\description{
ruf
}
