% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadate_cols.R
\name{add_metadate_cols}
\alias{add_metadate_cols}
\title{ruf}
\usage{
add_metadate_cols(df, time_var = "timestamp",
  time_format = "\%m/\%d/\%Y", time_origin = "1960-01-01")
}
\arguments{
\item{df}{class: dataframe}

\item{time_var}{class: string}

\item{time_format}{class: string}

\item{time_origin}{class: string}
}
\description{
ruf
}
\examples{
add_metadate_cols(df)
}
\keyword{add}
\keyword{date}
\keyword{features}
\keyword{timestamp}
\keyword{to}
