% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_obs.R
\name{avg_obs}
\alias{avg_obs}
\title{ruf}
\usage{
avg_obs(data, id_var = "id")
}
\arguments{
\item{data}{class: dataframe}

\item{time_var}{class: string}

\item{time_format}{class: string}
}
\description{
ruf
}
\examples{
avg_obs(data, id_var="id")
}
\keyword{add}
\keyword{date}
\keyword{features}
\keyword{timestamp}
\keyword{to}
