% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman_analysis.R
\name{bland_altman_analysis}
\alias{bland_altman_analysis}
\title{ruf}
\usage{
bland_altman_analysis(df, id_var = "id", rt_var = "response_time",
  assay_type_col = NA, assay1_name = NA, assay2_name = NA)
}
\arguments{
\item{df}{class: dataframe}

\item{id_var}{class: string}

\item{rt_var}{class: string}

\item{assay_type_col}{class: string}

\item{assay1_name}{class: string}

\item{assay2_name}{class: string}
}
\description{
ruf
}
\examples{
bland_altman_analysis(df, id_var="id", rt_var = "response_time", assay_type_col = "game_type", assay1_name = "native", assay2_name = "script")
}
\keyword{add}
\keyword{date}
\keyword{features}
\keyword{timestamp}
\keyword{to}
