% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman_plot.R
\name{bland_altman_plot}
\alias{bland_altman_plot}
\title{ruf}
\usage{
bland_altman_plot(
  df,
  diff_col = NA,
  mean_col = NA,
  sd_loa = 1.96,
  color_loa = "red",
  font_size = 18
)
}
\arguments{
\item{df}{class: dataframe}

\item{diff_col}{class: string}

\item{mean_col}{class: string}

\item{sd_loa}{class: numeric}

\item{color_loa}{class: string}

\item{font_size}{class: numeric}
}
\description{
ruf
}
\examples{
bland_altman_plot(df, diff_col = NA, mean_col = NA, sd_loa = 1.96, color_loa = "red", font_size=18)
}
