% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_reliability.R
\name{calc_reliability}
\alias{calc_reliability}
\title{ruf}
\usage{
calc_reliability(fit, n_obs = NA)
}
\arguments{
\item{fit}{class: lme4 / nlme model object}

\item{n_obs}{class: numeric}
}
\description{
ruf
}
\examples{
calc_reliability(fit, n_obs=14)
}
\keyword{assessment}
\keyword{data}
\keyword{data,}
\keyword{ecological}
\keyword{intensive}
\keyword{longitudinal}
\keyword{momentary}
