% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2condition_response_time_dataset.R
\name{generate_2condition_response_time_dataset}
\alias{generate_2condition_response_time_dataset}
\title{ruf}
\usage{
generate_2condition_response_time_dataset(n = 10, n_sessions = 1,
  rt_mean = 500, rt_sd = 200, cond_cats = c("native", "script"),
  rand_seed = 999)
}
\arguments{
\item{n}{class: numeric}

\item{n_sessions}{class: numeric}

\item{rt_mean}{class: numeric}

\item{rt_sd}{class: numeric}

\item{cond_cats}{class: vector}

\item{rand_seed}{class: numeric}

\item{df}{class: dataframe}
}
\description{
ruf
}
\examples{
generate_2condition_response_time_dataset(n = 20, n_sessions = 2, rt_mean = 500, rt_sd = 25, cond_Cats = c("native","script"))
}
