% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_typing_dataset.R
\name{generate_typing_dataset}
\alias{generate_typing_dataset}
\title{ruf}
\usage{
generate_typing_dataset(
  start_ts = 1563937206,
  start_ts_offset = 1000,
  stream_length = 100,
  rand_seed = 999
)
}
\arguments{
\item{start_ts}{class: numeric}

\item{start_ts_offset}{class: numeric}

\item{stream_length}{class: numeric}

\item{rand_seed}{class: numeric}
}
\description{
ruf
}
\examples{
generate_typing_dataset(start_ts = 1563937206, start_ts_offset=1000, stream_length=100, rand_seed = 999)
}
\keyword{data}
\keyword{processing}
