% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_googlescholar_yearly_article_count.R
\name{get_googlescholar_yearly_article_count}
\alias{get_googlescholar_yearly_article_count}
\title{ruf}
\usage{
get_googlescholar_yearly_article_count(
  search_term,
  search_years,
  sleep_times,
  search_language = "en"
)
}
\arguments{
\item{search_term}{class: string}

\item{search_years}{class: vector}

\item{sleep_times}{class: Vector}

\item{search_language}{class: string}
}
\description{
ruf
}
\examples{
get_googlescholar_yearly_article_count(search_term, search_years, sleep_times, search_language="en")
}
