% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sleep_times.R
\name{get_rand_sleep_times}
\alias{get_rand_sleep_times}
\title{ruf}
\usage{
get_rand_sleep_times(min = 0, max = 10, by = 0.1)
}
\arguments{
\item{min}{class: numeric}

\item{max}{class: numeric}

\item{by}{class: numeric}
}
\description{
ruf
}
\examples{
get_rand_sleep_times(min=0, max=10, by=0.1)
}
