% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time_travel_date_vector.R
\name{get_time_travel_vector}
\alias{get_time_travel_vector}
\title{ruf}
\usage{
get_time_travel_vector(start_date, horizon_days = 14, horizon_direction = "+")
}
\arguments{
\item{start_date}{class: POSIXlt}

\item{horizon_days}{class: numeric}

\item{horizon_direction}{class: string}
}
\description{
ruf
}
\examples{
get_time_travel_vector(start_date, horizon_days = 14, horizon_direction = "+")
}
\keyword{add}
\keyword{date}
\keyword{features}
\keyword{timestamp}
\keyword{to}
