% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trims.R
\name{get_trims}
\alias{get_trims}
\title{ruf}
\usage{
get_trims(.data, trim_lower = NA, trim_upper = NA, p = NA)
}
\arguments{
\item{.data}{class: data.frame}

\item{trim_lower}{class: boolean}

\item{trim_upper}{class: boolean}

\item{p}{class: numeric}
}
\description{
ruf
}
\examples{
read_any ("C:/filepath.csv")
}
