% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_reliability.R
\name{iterative_reliability}
\alias{iterative_reliability}
\title{ruf}
\usage{
iterative_reliability(
  data,
  varname,
  cluster_var = "studyday",
  id_var = "id",
  min_day = 0,
  max_day = 14,
  autocorrelated = T
)
}
\arguments{
\item{data}{class: data frame}

\item{varname}{class: string}

\item{cluster_var}{class: string}

\item{id_var}{class: string}

\item{min_day}{class: numeric}

\item{max_day}{class: numeric}

\item{autocorrelated}{class: boolean}
}
\description{
ruf
}
\examples{
iterative_reliability(data, "response_time", "studyday", min_day=0, max_day=14, autocorrelated=T)
}
\keyword{assessment}
\keyword{data}
\keyword{data,}
\keyword{ecological}
\keyword{intensive}
\keyword{longitudinal}
\keyword{momentary}
