% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm2pie.R
\name{lmm2pie}
\alias{lmm2pie}
\title{ruf}
\usage{
lmm2pie(fit, pkg = "lme4")
}
\arguments{
\item{fit}{class: model fit output object}

\item{pkg}{class: string}
}
\description{
ruf
}
\examples{
lmm2pie(fit, pkg="lme4")
}
\keyword{assessment}
\keyword{data}
\keyword{data,}
\keyword{ecological}
\keyword{intensive}
\keyword{longitudinal}
\keyword{momentary}
