% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_colnames.R
\name{make_tidy_colnames}
\alias{make_tidy_colnames}
\title{ruf}
\usage{
make_tidy_colnames(.data, unique = FALSE)
}
\arguments{
\item{.data}{class: data frame}

\item{unique}{class: boolean}
}
\description{
ruf
}
\examples{
make_tidy_colnames(data, unique = FALSE)
}
