% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_datetime.R
\name{make_tidy_datetime}
\alias{make_tidy_datetime}
\title{ruf}
\usage{
make_tidy_datetime(datetime = Sys.time(), timezone = "UTC")
}
\arguments{
\item{datetime}{class: POSIXct}

\item{timezone}{class: string}
}
\description{
ruf
}
\examples{
make_tidy_datetime(datetime=NA, timezone="UTC")
}
