% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tidy_datetime_filename.R
\name{make_tidy_datetime_filename}
\alias{make_tidy_datetime_filename}
\title{ruf}
\usage{
make_tidy_datetime_filename(datetime = Sys.time(), timezone = "UTC",
  prefix = "", suffix = "")
}
\arguments{
\item{datetime}{class: POSIXct}

\item{timezone}{class: string}

\item{prefix}{class: string}

\item{suffix}{class: string}
}
\description{
ruf
}
\examples{
make_tidy_datetime(datetime=NA, timezone="UTC")
}
