% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_dir_files.R
\name{merge_dir_files}
\alias{merge_dir_files}
\title{ruf}
\usage{
merge_dir_files(mypath, pattern, recursive = T, return_list = T)
}
\arguments{
\item{mypath}{class: string}

\item{pattern}{class: string}

\item{recursive}{class: boolean}

\item{return_list}{class: boolean}
}
\description{
ruf
}
\examples{
merge_dir_files(mypath, pattern, recursive=T, return_list = T)
}
