% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_obs_per_person.R
\name{n_obs_per_person}
\alias{n_obs_per_person}
\title{ruf}
\usage{
n_obs_per_person(data = NULL, id_var = NULL, expected_obs = NULL)
}
\arguments{
\item{data}{class: dataframe}

\item{id_var}{class: string}

\item{expected_obs}{class: numeric}
}
\description{
ruf
}
\examples{
n_obs_per_person(df)
}
\keyword{assessment}
\keyword{data}
\keyword{data,}
\keyword{ecological}
\keyword{intensive}
\keyword{longitudinal}
\keyword{momentary}
