% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_exec.R
\name{pause_exec}
\alias{pause_exec}
\title{ruf}
\usage{
pause_exec(prompt, r = "character")
}
\arguments{
\item{prompt}{class: string}

\item{r}{class: string}
}
\description{
ruf
}
\examples{
pause_exec(prompt, r="character")
}
