% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_googlescholar_citation_trend.R
\name{plot_googlescholar_citation_trend}
\alias{plot_googlescholar_citation_trend}
\title{ruf}
\usage{
plot_googlescholar_citation_trend(results, color = "#0077ff", size = 6,
  angle = 45)
}
\arguments{
\item{results}{class: tibble}

\item{color}{class: string}

\item{size}{class: numeric}

\item{angle}{class: numeric}
}
\description{
ruf
}
\examples{
plot_googlescholar_citation_trend(results, color="#0077ff", size=6, angle=45)
}
