% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_change.R
\name{prop_change}
\alias{prop_change}
\title{ruf}
\usage{
prop_change(old, new, rd = 2, perc = T, print_str = T)
}
\arguments{
\item{old}{class: numeric}

\item{new}{class: numeric}

\item{rd}{class: numeric}

\item{perc}{class: boolean}

\item{print_str}{class: boolean}
}
\description{
ruf
}
