% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_datetime.R
\name{rand_datetime}
\alias{rand_datetime}
\title{ruf}
\usage{
rand_datetime(start_date, end_date, size, tz = "America/New_York",
  rand_seed = 999, verbose = T, return_vector = F)
}
\arguments{
\item{start_date}{class: numeric}

\item{end_date}{class: numeric}

\item{size}{class: numeric}

\item{rand_seed}{class: numeric}
}
\description{
ruf
}
\examples{
rand_datetime(start_date, end_date, size, rand_seed)
}
\keyword{data}
\keyword{processing}
