% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruf-package.R
\docType{package}
\name{ruf}
\alias{ruf}
\title{Software to Implement Resource Utilization Function Estimation}
\description{
This package provides functionality 
to determine maximum
likelihood fits of the Resource Utilization Function based on a
Matern covariance function.
}
