% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_reliability.R
\name{stepwise_reliability}
\alias{stepwise_reliability}
\title{ruf}
\usage{
stepwise_reliability(data, varname, cluster_var = "studyday",
  id_var = "id", listen = c(1, 2), autocorrelated = T)
}
\arguments{
\item{data}{class: data frame}

\item{varname}{class: string}

\item{cluster_var}{class: string}

\item{id_var}{class: string}

\item{listen}{class: vector (which points of cluster_var do you want reliability for)}

\item{autocorrelated}{class: boolean}
}
\description{
ruf
}
\examples{
stepwise_reliability(data, "response_time", "studyday", min_day=0, max_day=14, autocorrelated=T)
}
\keyword{assessment}
\keyword{data}
\keyword{data,}
\keyword{ecological}
\keyword{intensive}
\keyword{longitudinal}
\keyword{momentary}
