% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpcr.R
\name{wpcr}
\alias{wpcr}
\title{ruf}
\usage{
wpcr(data = NULL, id_var = NULL, x = NULL, y = NULL,
  cov_fixed_effects = NULL, cov_random_effects = NULL)
}
\arguments{
\item{data}{class: dataframe}

\item{id_var}{class: string}

\item{x}{class: string}

\item{y}{class: string}

\item{cov_fixed_effects}{class: string, vector of strings}

\item{cov_random_effects}{class: string, vector of strings}

\item{id_var}{class: string}
}
\description{
ruf
}
\examples{
wpcr(data = data, x = "X", y = "Y", id_var = "ID", cov_fixed_effects = NULL, cov_random_effects = NULL)
}
\keyword{assessment}
\keyword{data}
\keyword{data,}
\keyword{ecological}
\keyword{intensive}
\keyword{longitudinal}
\keyword{momentary}
