% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_append_filename.R,
%   R/read_append_filename2.R
\name{read_append_filename}
\alias{read_append_filename}
\alias{read_append_filename_csv}
\title{ruf}
\usage{
read_append_filename(filename, verbose = F)

read_append_filename_csv(filename, verbose = F)
}
\arguments{
\item{filename}{class: string}

\item{verbose}{class: boolean}

\item{filename}{class: string}

\item{verbose}{class: boolean}
}
\description{
ruf

ruf
}
\examples{
read_append_filename(filename, verbose=F)
read_append_filename(filename, verbose=F)
}
